+++ app/pb_kits/playbook/pb_date_time_stacked/docs/_date_time_stacked_default.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { DateTimeStacked } from '../../'
+
+const DateTimeStackedDefault = () => (
+  <div>
+    <DateTimeStacked date={new Date} />
+  </div>
+)
+
+export default DateTimeStackedDefault