+++ app/pb_kits/playbook/pb_time_stacked/docs/index.js
@@ -1,0 +1,3 @@
+export { default as TimeStackedDefault } from './_time_stacked_default.jsx'
+export { default as TimeStackedDark } from './_time_stacked_dark.jsx'
+export { default as TimeStackedCaption } from './_time_stacked_caption.jsx'