+++ CHANGELOG.md
@@ -5,8 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
+## Unreleased
+
+## [4.13.0] 2020-4-24
+
+### Added
+- Added card header prop to Card Kit and added category colors ([#746](https://github.com/powerhome/playbook/pull/746)@christinaatai)
+- New Kits - Time Stacked & Date Time Stacked ([#745](https://github.com/powerhome/playbook/pull/745) @haydenrou)
+- Added table row side highlight prop ([#753](https://github.com/powerhome/playbook/pull/753) @christinaatai)
+
+### Fixed
+- Limit webpacker CPU & memory on local dev ([#749](https://github.com/powerhome/playbook/pull/749) @thestephenmarshall)
 - Milano branded deployment config ([#748](https://github.com/powerhome/playbook/pull/748) @benlangfeld)
 
 ## [4.12.0] 2020-4-17